+++ processor/src/test/resources/expected/com/example/arez/React4j_ComponentWithDependency.java
@@ -9,6 +9,7 @@ import javax.annotation.Generated;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import jsinterop.annotations.JsConstructor;
+import jsinterop.annotations.JsPackage;
 import jsinterop.annotations.JsType;
 import jsinterop.base.Js;
 import jsinterop.base.JsPropertyMap;
@@ -79,7 +80,9 @@ abstract class React4j_ComponentWithDependency extends ComponentWithDependency {
   }
 
   @JsType(
-      isNative = true
+      isNative = true,
+      namespace = JsPackage.GLOBAL,
+      name = "?"
   )
   interface Lifecycle {
     void componentDidMount();