+++ processor/src/test/resources/expected/com/example/arez/React4j_ComponentWithNoAutoActionCallback.java
@@ -7,6 +7,7 @@ import javax.annotation.Generated;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import jsinterop.annotations.JsConstructor;
+import jsinterop.annotations.JsPackage;
 import jsinterop.annotations.JsType;
 import jsinterop.base.Js;
 import jsinterop.base.JsPropertyMap;
@@ -49,7 +50,9 @@ abstract class React4j_ComponentWithNoAutoActionCallback extends ComponentWithNo
   }
 
   @JsType(
-      isNative = true
+      isNative = true,
+      namespace = JsPackage.GLOBAL,
+      name = "?"
   )
   interface Lifecycle {
     void componentDidMount();