+++ processor/src/test/resources/expected/com/example/arez/React4j_ComponentWithProp.java
@@ -8,6 +8,7 @@ import javax.annotation.Generated;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import jsinterop.annotations.JsConstructor;
+import jsinterop.annotations.JsPackage;
 import jsinterop.annotations.JsType;
 import jsinterop.base.Js;
 import jsinterop.base.JsPropertyMap;
@@ -53,7 +54,9 @@ abstract class React4j_ComponentWithProp extends ComponentWithProp {
   }
 
   @JsType(
-      isNative = true
+      isNative = true,
+      namespace = JsPackage.GLOBAL,
+      name = "?"
   )
   interface Lifecycle {
     void componentDidMount();