+++ processor/src/test/resources/expected/com/example/arez/React4j_OverridingComponentDidUpdateComponent.java
@@ -6,6 +6,7 @@ import javax.annotation.Generated;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import jsinterop.annotations.JsConstructor;
+import jsinterop.annotations.JsPackage;
 import jsinterop.annotations.JsType;
 import jsinterop.base.Js;
 import jsinterop.base.JsPropertyMap;
@@ -35,7 +36,9 @@ abstract class React4j_OverridingComponentDidUpdateComponent extends OverridingC
   }
 
   @JsType(
-      isNative = true
+      isNative = true,
+      namespace = JsPackage.GLOBAL,
+      name = "?"
   )
   interface Lifecycle {
     void componentDidMount();