+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * Upgrade the `org.realityforge.braincheck:braincheck:jar` dependency to `1.10.0` for improved
   compatibility with J2CL with respect to compile-time constants.
+* **\[processor\]** Change the jsinterop typing of the `Lifecycle` interface applied to native React4j
+  components from `@JsType(isNative = true)` to `@JsType(isNative = true,namespace = JsPackage.GLOBAL,name = "?")`
+  as that seems to be more inline with the code expected by J2CL or more precisely the closure compiler.
 
 ### [v0.80](https://github.com/react4j/react4j/tree/v0.80) (2018-06-28)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.79...v0.80)